pyproject.toml
[project]
name = "insightforge"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.1.6",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
]
